[Chore] Restructured files and enhanced ServerService for testing
