+++ src/index.js
@@ -1,5 +1,11 @@
-import ServerService from "./services/serverService.js";
+import ServerService from "./services/server/serverService.js";
 
 // Start Server Service
 const server = new ServerService();
-server.start();
+server.start()
+    .on('listening', () => {
+        console.log("Server initialized!");
+    })
+    .on('error', (error) => {
+        console.log(`Server failed to initialize: ${error.message}`)
+    });
